refactor(tvc): seal operator backends behind a Signer port
