+++ tvc/src/commands/app/init.rs
@@ -2,6 +2,7 @@
 
 use crate::config::app::AppConfig;
 use crate::config::turnkey::{Config, StoredQosOperatorKey};
+use crate::local_operator_key::select_local_operator;
 use crate::outcome::Outcome;
 use crate::output::StdCtx;
 use crate::prompts::{bail_interactive_conflicts_with_non_interactive, ensure_stdin_is_tty};
@@ -109,10 +110,8 @@ async fn load_operator_public_key() -> Option<String> {
     let config = Config::load().await.ok()?;
 
     // Get active org config
-    let (org_id, org_config) = config.active_org_config()?;
-    let local = org_config
-        .select_local_record(&config.display_name(org_id))
-        .ok()?;
+    let (_, org_config) = config.active_org_config()?;
+    let (_, local) = select_local_operator(org_config).ok()?;
     let operator_key = StoredQosOperatorKey::load(&local.key_path).await.ok()??;
     Some(operator_key.public_key.to_string())
 }
