+++ tvc/src/commands/keys/backup_operator_key.rs
@@ -4,6 +4,7 @@
 use crate::{
     commands::{Run, login::resolve_org_query},
     config::turnkey::{Config, OrgQuery, QosOperatorPublicKey, StoredQosOperatorKey},
+    local_operator_key::select_local_operator,
     outcome::Outcome,
     output::StdCtx,
     prompts::{self, error_required_in_non_interactive},
@@ -56,9 +57,9 @@ impl Run for Args {
                     .ok_or_else(|| anyhow!("No active organization. Run `tvc login` first."))
             })?;
 
-        let source = &org_config
-            .select_local_record(&resolved.to_string())?
-            .key_path;
+        let (_, local) =
+            select_local_operator(org_config).with_context(|| format!("org '{resolved}'"))?;
+        let source = &local.key_path;
 
         let destination: PathBuf = self.output.map(Ok).unwrap_or_else(|| {
             prompts::text(
